Rename _utils directory to mocks
